front/requirements.txt
streamlit
numpy<2.0.0
opencv-python-headless
nest-asyncio
lark
pysqlite3-binary
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-chroma
chromadb
pandas
protobuf<5
ultralytics
packaging>=23.0